+++ pygeotemporal_parsers/parse/parse_new_datapoints.py
@@ -71,6 +71,9 @@ def update_sensors_stats(sensor_client, sensor_names):
     """Update all sensors stats after uploads"""
 
     for sensor_name in sensor_names:
+        if '&' in sensor_name:
+            sensor_name = str(sensor_name.replace('&', '%26'))
+
         sensor_raw = sensor_client.sensor_get_by_name(sensor_name)
         sensor = sensor_raw.json()
         if len(sensor) == 1:
@@ -106,6 +109,8 @@ def parse_data(timestamp, config, sensor_names, parameters, datafile,
 
     for sensor_name in sensor_names:
 
+        if '&' in sensor_name:
+            sensor_name = str(sensor_name.replace('&', '%26'))
         print 'sensor_name is ' + sensor_name
 
         # Get Sensor information for the Sensor Name
@@ -114,7 +119,6 @@ def parse_data(timestamp, config, sensor_names, parameters, datafile,
         sensor_id = parse_sensor['id']
 
         # Get Stream information for the Sensor Name
-        sensor_name = parse_sensor['name']
         stream_raw = stream_client.stream_get_by_name_json(sensor_name)
         parse_stream = stream_raw[0]
         stream_id = parse_stream['id']